Resolves HELIO-4051 Alum does not have access but the access icon is unlocked green
