+++ app/views/fulcrum/_dashboard.html.erb
@@ -49,7 +49,7 @@
     </li>
     <li>
       Licenses
-      <% Incognito.sudo_actor_licenses(current_actor).each do |license| %>
+      <% current_actor.licenses.each do |license| %>
         <ul>
           <li><%= link_to " #{license.label} #{license.license_affiliations.map(&:affiliation).join(', ')} access to #{license.product.identifier}", greensub_license_path(license) %></li>
         </ul>
@@ -58,7 +58,7 @@
     </li>
     <li>
       Products
-      <% Incognito.sudo_actor_products(current_actor).each do |product| %>
+      <% current_actor.products.each do |product| %>
         <ul>
           <li><%= link_to "#{product.identifier} - #{product.name}", product %></li>
         </ul>
